+++ assets/training/model_management/environments/foundation-model-inference/context/src/managed_inference.py
@@ -4,13 +4,17 @@
 """This module provides the MIRPayload class that codifies the payload that is received in the scoring script."""
 
 import json
+import os
 from dataclasses import dataclass
 from typing import Any, Dict, List, Tuple, Union
 from configs import SerializableDataClass
 from constants import TaskType
 from logging_config import configure_logger
+from transformers import AutoTokenizer
 
 logger = configure_logger(__name__)
+
+DEFAULT_MLFLOW_MODEL_PATH = "mlflow_model_folder/data/model"
 
 
 @dataclass
@@ -29,25 +33,21 @@ class MIRPayload(SerializableDataClass):
 
     def convert_query_to_list(self) -> None:
         """
-        Convert the query prompts into a list.
+        Convert the query parameter into a list.
 
-        FMScore.run expects a list of prompts.
-        In the case of chat completion, a single string is produced and needs to be
-        put inside of a list.
+        FMScore.run expects a list of prompts. In the case of chat completion, a single string
+        is produced and needs to be put inside of a list.
         """
         if not isinstance(self.query, list):
             self.query = [self.query]
 
     def update_params(self, new_params: Dict) -> None:
-        """Update the params with the given dictionary."""
+        """Update current parameters to the new parameters the MIRPayload should have."""
         self.params = new_params
 
 
 def get_processed_input_data_for_chat_completion(dialog: List[str]) -> str:
     r"""Process chat completion input request.
-
-    Taken from:
-    https://github.com/facebookresearch/llama/blob/main/llama/generation.py
 
     example input:
     [
@@ -69,55 +69,12 @@ def get_processed_input_data_for_chat_completion(dialog: List[str]) -> str:
     As of 2021, the Burj Khalifa in Dubai\n
     [INST]and in Africa?[/INST]"
     """
-    SPECIAL_TAGS = ["[INST]", "[/INST]", "<<SYS>>", "<</SYS>>"]
-    UNSAFE_ERROR = "Error: special tags are not allowed as part of the prompt."
-
-    def process_dialog(messages) -> Tuple[str, List[Tuple[str, str]], str]:
-        system_prompt = ""
-        user_assistant_messages = []  # list of (user, assistant) messages
-        user_message = None  # current user message being processed
-        last_user_message = None  # user prompt for which response is needed
-
-        unsafe_request = any([tag in msg["content"] for tag in SPECIAL_TAGS for msg in messages])
-        if unsafe_request:
-            raise Exception(UNSAFE_ERROR)
-
-        for i, message in enumerate(messages):
-            role = message["role"]
-            content = message["content"]
-
-            if role == "system" and i == 0:
-                system_prompt = content
-            elif role == "user":
-                if i == len(messages) - 1:
-                    last_user_message = content
-                else:
-                    user_message = content
-            elif role == "assistant" and user_message is not None:
-                user_assistant_messages.append((user_message, content))
-                user_message = None
-
-        return system_prompt, user_assistant_messages, last_user_message
-
-    # ref: https://huggingface.co/spaces/huggingface-projects/\
-    # llama-2-7b-chat/blob/main/model.py
-    def format_chat_conv(message: str, chat_history: List[Tuple[str, str]], system_prompt: str) -> str:
-        texts = [f'<s>[INST] <<SYS>>\n{system_prompt}\n<</SYS>>\n\n'] \
-            if system_prompt != "" \
-            else ['<s>[INST] ']
-        # The first user input is _not_ stripped
-        do_strip = False
-        for user_input, response in chat_history:
-            user_input = user_input.strip() if do_strip else user_input
-            do_strip = True
-            texts.append(
-                f'{user_input} [/INST] {response.strip()} </s><s>[INST] ')
-        message = message.strip() if do_strip else message
-        texts.append(f'{message} [/INST]')
-        return ''.join(texts)
-
-    sys_prompt, user_assistant_msgs, message = process_dialog(dialog)
-    chat_conv = format_chat_conv(message, user_assistant_msgs, sys_prompt)
+    # get path to model folder
+    model_path = str(os.path.join(os.getenv("AZUREML_MODEL_DIR", ""), DEFAULT_MLFLOW_MODEL_PATH))
+    # use tokenizer defined in tokenizer_config
+    tokenizer = AutoTokenizer.from_pretrained(model_path, use_default_system_prompt=False)
+    # apply template to format chat conversation
+    chat_conv = tokenizer.apply_chat_template(dialog, tokenize=False)
     return chat_conv