+++ assets/training/model_management/src/azureml/model/mgmt/processors/pyfunc/config.py
@@ -44,6 +44,15 @@ class SupportedTasks(_CustomEnum):
     # LLaVA task
     IMAGE_TEXT_TO_TEXT = "image-text-to-text"
 
+    # mask generation
+    MASK_GENERATION = "mask-generation"
+
+    # AutoML Tasks
+    IMAGE_CLASSIFICATION = "image-classification"
+    IMAGE_CLASSIFICATION_MULTILABEL = "image-classification-multilabel"
+    IMAGE_OBJECT_DETECTION = "image-object-detection"
+    IMAGE_INSTANCE_SEGMENTATION = "image-instance-segmentation"
+
 
 class SupportedTextToImageModelFamily(_CustomEnum):
     """Supported text to image models."""